Combine build/docker steps.
